refactor fine-tuning into phases
